Modify source code
Add a error handling for express routing. This error handling will catch
jwt exception caused by jwt-express, and catch a custom error
RequestError and return the coresponding status code, error code and
message.

Change test reporter to be spec. The process.env for testing is modified
in gulpfile, not in test file.